+++ packages/ui-shared/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sharelyai/ui-shared
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @sharelyai/services@0.0.4
+
 ## 0.0.2
 
 ### Patch Changes